process errors directly in queue
